Verigames changes for refinement variables and BallSizeTestConstraints (IfNodeTests).
Now handles BallSizeTestConstraints though BallSizeTestConstraintHandler. The small and large ports of the BSTest are added to the board along with other RefinementVariable, but instead of introducing them to the board with a START_* they are started with a CONNECT to another CONNECT.  The BallSizeTestConstraintHandler then attaches the starting CONNECTs to the output of the BallSizeTest intersection.  The CONNECT's created to introduce the BSTest refinement variables are kept in a new hash, boardNBSVariableToStart which is used to look them up later.

RefinementVariables are now started with START_PIPE_DEPENDENT_BALL. This was because RefinementVariables no longer have subtype constraints flowing into them, they instead have an equality constraint (through linking) to the right-hand-side value that was being assigned.

NNinfTransfer.java creates BallSizeTestConstraints and caches them. The constraint's are added to the constraint manager by GameVisitor visitBinary. Handles cycles.

TrustedGameSolver.scala needed modifications to handle RefinementVariables.

Level.java now sets all linked chutes to the same width, defaulting to narrow if no chutes have uneditable. If there are conflicting uneditable chutes an exception is raised. If one chutes is uneditable, then all are set to uneditable.

The DotPrinter.java has been updated to add more useful labels to the output.

Misc Changes to verigamesTest.py to allow running individual tests based on a regex match, better output, more resiliency to missing gold files, and automatic generation/updating of gold files.