Added the option to preserve IDs when loading a world
This is an important precondition for some upcoming optimizer
improvements. We want to be able to map specific edges to specific other
edges (e.g. for mapping buzzsaw placement) but this needs to be done
per edge rather than per variable. Thus we need consistent edge IDs
during the world export/import process.