Improve the performance and correctness of edge set conversion
Previously, performance would be very very bad for large edge sets since
we were storing the set of edges as a hash key. The new approach works
only on the var ID level (no edges) and is much more efficient.

In addition, this fixes a bug where if all the edges with a particular
var ID have been eliminated, a null pointer exception could be thrown.